Widen skinned joint indices to u32, matching the shader's uint4
The skinning shader declares its joint indices as uint4 -- 32-bit -- while the
vertex format said USHORT4, leaving the fetch to widen 16 bits to 32. D3D12 did
that correctly and Vulkan did not, on the same data from the same file, which
is why a character skinned cleanly on Windows and threw a vertex across the
room on Linux.

The consequence was quiet rather than obvious: an out-of-range read of the
joint palette is defined on D3D12, returning zero, and undefined on Vulkan. So
a joint index that widened wrongly was absorbed by one backend and turned into
a garbage matrix by the other.

Isolated by elimination: the artifact survived the bind pose, so no clip,
layer or palette was involved; it survived hiding the grid and the blocks, so
it was the character; forcing the joints and weights to constants removed it;
and splitting that in two showed the weights were fine and the joints were not.
Every byte in the buffer was verified correct beforehand -- weights summing to
1.0000, no index above 64 against a 66-joint skin, and all 171,809 zero-weight
slots naming joint 0 -- so the data was never the problem, only the fetch.

The vertex grows from 56 bytes to 64, which as a side effect puts joints at
offset 32 and weights at 48, both 16-byte aligned, where the old stride was
divisible by no useful alignment. The shader is unchanged, so the committed
.spv and .dxil stay in step.

+++ matchbox/model_load.odin
@@ -340,9 +340,9 @@ skinned_primitive_part :: proc(
 		vertex pinned to the root joint is a visible seam rather than a
 		character stretched across the map.
 	*/
-	joint_limit := u16(MAX_JOINTS)
+	joint_limit := u32(MAX_JOINTS)
 	if skin >= 0 && skin < len(data.skins) {
-		joint_limit = u16(min(len(data.skins[skin].joints), MAX_JOINTS))
+		joint_limit = u32(min(len(data.skins[skin].joints), MAX_JOINTS))
 	}
 
 	out_of_range := 0
